refactor(swap): extract calculateFees() pipeline into SwapQuoteOrchestrator (#4735)
Move the reactive quote/fee orchestration out of SwapFormViewModel into a
new SwapQuoteOrchestrator collaborator behind a single flow, leaving
calculateFees() a thin delegator. The pipeline is relocated verbatim over a
shared SwapQuoteContext so every state-write site and its ordering — the
UTXO plan-fee race, the isLoading spinner timing (#4712), and the
isSwapDisabled gating — is preserved exactly.

SwapFormViewModel drops from 1145 to 622 lines and sheds six pipeline-only
dependencies (swapQuoteRepository, appCurrencyRepository, referralRepository,
getDiscountBpsUseCase, swapDiscountChecker, ioDispatcher).
